fix(predictor): ruff lint + cubic P1/P2 review fixes
CI lint (ruff):
- sort imports (I001), drop unused math / torch.nn (F401)

P1 (correctness):
- pad expert indices with -1, not 0 (0 is a valid expert id, so 0
  padding corrupts labels when K varies). Filter negatives in
  build_target_multilabel + every eval path, and count real experts
  (not padded K) in temporal/top-k/per-layer hit-rate denominators.
- guard empty train/test split (exit cleanly instead of crashing).
- MoeRoutingCollector::record now checks every fwrite; a short write
  closes the file and stops, instead of emitting misaligned/corrupt
  training data with an inflated sample count.

P2:
- int64_t sample counter (+ %lld).
- moe_hybrid_routing_stats: cap top-10/30/60 to total when
  n_expert < bucket size (was printing 0%).
- --freq/--collect-routing force routing-stats env vars when unset or
  empty (overwrite=0 left an empty value → silent no-op).
- set_routing_collector returns bool (capability check); caller warns +
  closes collector when the backend can't collect; lifetime contract
  documented.
- dedup the per-layer ffn_post D2H copy across all three decode paths
  when --collect-routing and cold compute are both active.
- correct --hidden-dim/--num-experts/--num-layers help defaults.
- validate --threshold is in (0,1); handle empty/truncated routing files.

+++ server/src/common/moe_routing_collector.h
@@ -38,12 +38,12 @@ public:
     // Close and flush the output file. Prints summary to stderr.
     void close();
 
-    bool is_open() const { return fd_ != nullptr; }
-    int  sample_count() const { return samples_; }
+    bool    is_open() const { return fd_ != nullptr; }
+    int64_t sample_count() const { return samples_; }
 
 private:
     std::FILE * fd_ = nullptr;
-    int samples_ = 0;
+    int64_t samples_ = 0;  // wide: collection runs can exceed 2^31 samples
     std::mutex mu_;
 };
